A3: remove double from per-sample and inner-loop paths
The first hardware target is RV32IMAFC: single-precision FPU, no D extension,
so every double operation is a software-emulated library call. double was used
per source sample in three accumulators and throughout both autocorrelations.

Step 1, the autocorrelation accumulators. numerator, left_square, right_square,
score, best_scores[], the phase score, the candidate score ratio, the
confidence separation and apta_s6_window_t.score are float; sqrt became sqrtf.
The degenerate-energy guard moved from 1e-18 to 1e-12f, still far below any
real signal but meaningful in single precision.

Step 2, the per-sample accumulators.

apta_internal_onset_bin_t.sum_absolute is uint32_t, scaled by 2^15. The struct
is shared by S4 and S6 -- the work order asks whether it is, and it is -- so
this is one change covering both engines, and the bound comes from S6's 2048
samples per bin rather than S4's 256: 2048 * 32768 = 2^26, a 64x margin.

apta_internal_waveform_accumulator_t.sum_squares is uint64_t, scaled by 2^23,
not the 2^15 the work order's sketch implies. That difference is the point.
With 2^15 the published rms of some columns moved by one count out of 65535 --
a silent change to published data that no test in this repository detects,
because there are no stored golden column values; apta.waveform.determinism
compares two chunkings against each other, not against a reference. I measured
it by diffing column output against the A2 build. uint64_t has room for a much
finer scale: at 2^23 a sample decoded from 16-bit PCM, exactly k/32768, becomes
k * 2^8 with no loss at all. Published minimum, maximum, rms and flags are now
bit-identical to A2.

The work order's section 5.1 table misses the earliest per-sample double path.
apta_normalize_s16 / _s24 / _s32 in apta_waveform_input.c and their twins in
apta_waveform_detail_replay.c computed (float)((double)value / 32768.0) once
per source sample per channel, before anything else touches the sample. These
are now float divisions, bit-identically so: for 16- and 24-bit input both
operand and divisor are exactly representable in float, so IEEE 754 gives the
same correctly rounded result. apta_normalize_s32's positive branch keeps
double, because a 32-bit magnitude exceeds float's 24-bit mantissa; it is the
only per-sample double left and is commented as such.

Clamping uses fminf(fabsf(sample), 1.0f). apta_read_channel_sample() already
guarantees a finite value in [-1, 1], but the float-to-integer conversion must
not depend on a caller invariant for defined behaviour. Branchless also happens
to be safer: with an if, a NaN would pass through to an undefined integer
conversion; fminf returns the non-NaN operand.

Cost probe, us per call, 5 min @ 44.1 kHz:

                                     after A2   after A3
  overview                              155.4      152.5
  overview+confidence                   376.3      446.7
  +BPM                                  368.5      448.6
  +local grid                           371.2      447.5
  +global grid                          421.9      583.9
  +dynamic tempo                        418.6      520.4
  +detail+locking (full)                465.6      538.7

A3 costs about 20% on this host, which is expected and not a defect: on x86-64
float and double are both native, so the substitution buys nothing while the
integer scaling and conversions cost something. The benefit exists only where
double is emulated. Two caveats: run-to-run variance on the S6 rows is around
15% -- an earlier run of the same binary gave 500.9 for +global grid against
583.9 here -- and the work order's overall "every row under 500 us" target,
which A2 met, is no longer cleanly met, with the S6 rows straddling the line
inside that noise. A branchless clamp was tried as a mitigation and made no
measurable difference; it was kept for the NaN property, not for speed.

Acceptance criterion 4, that the ESP-IDF component-size report does not grow,
is NOT verified. The x86-64 static library grew from 131,695 to 134,005 bytes
of text, but that is the wrong measurement: on RV32 the change should remove
soft-float support routines rather than add code. No RISC-V toolchain or
ESP-IDF installation was available, so this is left to espidf.yml in CI and
should be checked there before the work is considered complete.

Tests: 70/70 default (Release, APTA_WARNINGS_AS_ERRORS=ON, no warnings), 60/60
core-only, 70/70 under ASan/UBSan with no diagnostics -- float-cast-overflow
being the check that matters here. The three ESP-IDF memory profiles are
unchanged. No test expectations were modified; apta.s4.tempo_vectors and
apta.waveform.determinism pass unmodified.

apta_pool_round_ties_even() and the double arithmetic feeding canonical
serialization are untouched, as the work order requires. The MSVC
long-double alignment shim is untouched; it is used only for alignof.

+++ src/waveform/apta_waveform_detail_replay.c
@@ -191,12 +191,14 @@ apta_status_t apta_internal_detail_next_pcm_request(
     return APTA_STATUS_OK;
 }
 
+/* A3: per-sample conversion. See apta_normalize_s16() in
+ * apta_waveform_input.c for why float is bit-identical here. */
 static float apta_detail_normalize_s16(int16_t value)
 {
     if (value < 0) {
-        return (float)((double)value / 32768.0);
+        return (float)value / 32768.0f;
     }
-    return value == 0 ? 0.0f : (float)((double)value / 32767.0);
+    return value == 0 ? 0.0f : (float)value / 32767.0f;
 }
 
 static float apta_detail_normalize_s24(const uint8_t *bytes)
@@ -209,16 +211,17 @@ static float apta_detail_normalize_s24(const uint8_t *bytes)
         value |= (int32_t)0xFF000000;
     }
     if (value < 0) {
-        return (float)((double)value / 8388608.0);
+        return (float)value / 8388608.0f;
     }
-    return value == 0 ? 0.0f : (float)((double)value / 8388607.0);
+    return value == 0 ? 0.0f : (float)value / 8388607.0f;
 }
 
 static float apta_detail_normalize_s32(int32_t value)
 {
     if (value < 0) {
-        return (float)((double)value / 2147483648.0);
+        return (float)value / 2147483648.0f;
     }
+    /* Retained in double: see apta_normalize_s32() in apta_waveform_input.c. */
     return value == 0 ? 0.0f : (float)((double)value / 2147483647.0);
 }
